requirements.txt
alabaster==0.7.12
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
codecov==2.1.3
coverage==5.1
cycler==0.10.0
decorator==4.4.2
Django==3.0.6
docutils==0.16
Flask==1.1.2
Flask-Session==0.3.1
google==2.0.3
idna==2.9
image==1.5.30
imagesize==1.2.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.2.0
msgpack-python==0.5.6
numpy==1.18.3
oleg-flask-sessions==0.2.5
olegdb-python==0.1.6
packaging==20.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==7.1.1
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.9.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
